// test/types.ts
export interface ShipEquipment {
  mainWeapon: number;
  armor: number;
  shields: number;
  special: number;
}

export interface ShipTraits {
  serialNumber: bigint;
  colors: ShipColors;
  variant: number;
  accuracy: number;
  hull: number;
  speed: number;
}

export interface ShipColors {
  h1: number;
  s1: number;
  l1: number;
  h2: number;
  s2: number;
  l2: number;
}

export interface ShipData {
  shipsDestroyed: number;
  costsVersion: number;
  cost: number;
  shiny: boolean;
  constructed: boolean;
  inFleet: boolean;
  isFreeShip: boolean;
  modified: number;
  timestampDestroyed: bigint;
}

export type ShipTuple = [
  string, // name
  bigint, // id
  ShipEquipment, // equipment
  ShipTraits, // traits
  ShipData, // shipData
  Address, // owner
];
// ...
export interface Attributes {
  version: number;
  range: number;
  gunDamage: number;
  hullPoints: number;
  maxHullPoints: number;
  movement: number;
  damageReduction: number;
  reactorCriticalTimer: number;
  statusEffects: number[];
}

export interface GameData {
  gameId: bigint;
  lobbyId: bigint;
  creator: Address;
  joiner: Address;
  creatorFleetId: bigint;
  joinerFleetId: bigint;
  creatorGoesFirst: boolean;
  startedAt: bigint;
  currentTurn: Address;
}

export interface GameMetadata {
  gameId: bigint;
  lobbyId: bigint;
  creator: Address;
  joiner: Address;
  creatorFleetId: bigint;
  joinerFleetId: bigint;
  creatorGoesFirst: boolean;
  startedAt: bigint;
  winner: Address;
  orchestrator: Address;
}

export interface GameTurnState {
  currentTurn: Address;
  turnTime: bigint;
  turnStartTime: bigint;
  currentRound: bigint;
}

export interface GameGridDimensions {
  gridWidth: number;
  gridHeight: number;
}
// ...
export interface Position {
  row: number; // Row position (0 to gridHeight-1)
  col: number; // Column position (0 to gridWidth-1)
}

export interface ShipPosition {
  shipId: bigint;
  position: Position;
  isCreator: boolean;
}

export interface LastMove {
  shipId: bigint;
  oldRow: number;
  oldCol: number;
  newRow: number;
  newCol: number;
  actionType: ActionType;
  targetShipId: bigint;
  timestamp: bigint; // When the move was made
}

export interface GameDataView {
  metadata: GameMetadata;
  turnState: GameTurnState;
  gridDimensions: GameGridDimensions;
  maxScore: bigint; // Maximum score needed to win the game
  creatorScore: bigint; // Current score of the creator player
  joinerScore: bigint; // Current score of the joiner player
  lastMove: LastMove; // Most recent move in the game
  shipAttributes: readonly Attributes[]; // Combined array of all ship attributes indexed by ship ID
  shipPositions: readonly ShipPosition[]; // All ship positions on the grid
  shipIds: readonly bigint[];
  creatorActiveShipIds: readonly bigint[];
  joinerActiveShipIds: readonly bigint[];
  creatorMovedShipIds: readonly bigint[]; // Creator ships that have moved this round
  joinerMovedShipIds: readonly bigint[]; // Joiner ships that have moved this round
}

export enum ActionType {
  Pass,
  Shoot,
  Retreat,
  Assist,
  Special,
  FactionAbility,
}

export enum MapMode {
  PvP,
  PvE,
  Both,
}